Event/pyproject.toml
[tool.poetry]
name = "event"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
django-filter = "^23.2"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
method-decorator = "^0.1.3"
djangorestframework = "^3.14.0"
django-oauth-toolkit = "^2.3.0"
djoser = "^2.2.0"
serializers = "^0.2.4"
status = "^0.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
